Add validations and handle exceptions (#7)
* Validations and exception handling for users

* Validations and exception handling for articles

* Validations and exception handling for orders

+++ WebshopServer/Services/ArticleService.cs
@@ -1,9 +1,11 @@
 ﻿using AutoMapper;
+using EntityFramework.Exceptions.Common;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using WebshopServer.Dtos;
+using WebshopServer.Exceptions;
 using WebshopServer.Infrastructure;
 using WebshopServer.Interfaces;
 using WebshopServer.Models;
@@ -28,35 +30,87 @@ namespace WebshopServer.Services
 
         public ArticleDto GetArticleById(long id)
         {
-            return _mapper.Map<ArticleDto>(_dbContext.Articles.Find(id));
+            ArticleDto article = _mapper.Map<ArticleDto>(_dbContext.Articles.Find(id));
+
+            if (article == null)
+            {
+                throw new ResourceNotFoundException("Article with specified id doesn't exist!");
+            }
+
+            return article;
         }
 
-        public ArticleDto CreateArticle(ArticleDto articleDto)
+        public ArticleDto CreateArticle(ArticleDto articleDto, long userId)
         {
             Article article = _mapper.Map<Article>(articleDto);
+            article.SellerId = userId;
 
             _dbContext.Articles.Add(article);
-            _dbContext.SaveChanges();
+
+            try
+            {
+                _dbContext.SaveChanges();
+            }
+            catch (CannotInsertNullException)
+            {
+                throw new InvalidFieldsException("One of more fields are missing!");
+            }
+            catch (Exception)
+            {
+                throw;
+            }
 
             return _mapper.Map<ArticleDto>(article);
         }
 
-        public ArticleDto UpdateArticle(long id, ArticleDto articleDto)
+        public ArticleDto UpdateArticle(long id, ArticleDto articleDto, long userId)
         {
             Article article = _dbContext.Articles.Find(id);
+
+            if (article == null)
+            {
+                throw new ResourceNotFoundException("Article with specified id doesn't exist!");
+            }
+
+            if (article.SellerId != userId)
+            {
+                throw new ForbiddenActionException("Sellers can only modify their own articles!");
+            }
+
             article.Name = articleDto.Name;
             article.Price = articleDto.Price;
             article.Quantity = articleDto.Quantity;
             article.Description = articleDto.Description;
 
-            _dbContext.SaveChanges();
+            try
+            {
+                _dbContext.SaveChanges();
+            }
+            catch (CannotInsertNullException)
+            {
+                throw new InvalidFieldsException("One of more fields are missing!");
+            }
+            catch (Exception)
+            {
+                throw;
+            }
 
             return _mapper.Map<ArticleDto>(article);
         }
 
-        public void DeleteArticle(long id)
+        public void DeleteArticle(long id, long userId)
         {
             Article article = _dbContext.Articles.Find(id);
+
+            if (article == null)
+            {
+                throw new ResourceNotFoundException("Article with specified id doesn't exist!");
+            }
+            
+            if (article.SellerId != userId)
+            {
+                throw new ForbiddenActionException("Sellers can only delete their own articles!");
+            }
 
             _dbContext.Articles.Remove(article);
             _dbContext.SaveChanges();
